+++ R/reporting.R
@@ -2,14 +2,13 @@
 #' @description Generate a QC performance report for a Olink HT
 #'  project.
 #'
+#' @param olink_assay
 #' @param project_name title to give the report
 #' @param correction_procedure description to use in
 #'  the report for the batch correction procedure
-#' @param sdp_directory path to the directory that
-#'  includes the project `Level_1` and `Level_2`
-#'  outputs.
 #' @param output_dir path to where the report should be output
 #' @param report_name name to give the output report
+#' @param template_file path to Quarto qmd file to use for report.
 #'
 #' @returns Nothing, but writes a PDF named `report_name` to `output_dir`
 #'
@@ -20,35 +19,27 @@
 #' @examples
 #' \dontrun{
 #' generate_performance_report(
-#'    project_name = "Volcano Science Project",
-#'    correction_procedure = "median",
-#'    sdp_directory = "/home/user/volcano_project/sdp",
+#'    olink_assay = here::here("olink_assay.RDS"),
+#'    project_name = "Science Fair Project",
+#'    correction_procedure = "Median correction",
 #'    output_dir = "/home/user/reports",
 #'    report_name = "assay_performance_report.pdf"
 #'  )
 #' }
 generate_performance_report <- function(
-  project_name,
-  correction_procedure,
-  sdp_directory,
-  output_dir,
-  report_name = "assay_performance_report.pdf"
+  olink_assay,
+  project_name = "Science Project",
+  correction_procedure = "median",
+  output_dir = here::here(),
+  report_name = "assay_performance_report.pdf",
+  template_file = NULL
 ) {
-  template_file <- system.file(
-    "qmd",
-    "performance_report_template.qmd",
-    package = "olinkqc"
-  )
-
-  level_1_path <- file.path(
-    sdp_directory,
-    "Level_1"
-  )
-
-  level_2_path <- file.path(
-    sdp_directory,
-    "Level_2"
-  )
+  template_file <- template_file %||%
+    system.file(
+      "qmd",
+      "performance_report_template.qmd",
+      package = "OlinkAssay"
+    )
 
   tmp_path <- withr::local_tempfile()
 
@@ -65,8 +56,7 @@ generate_performance_report <- function(
     execute_params = list(
       project_name = project_name,
       correction_procedure = correction_procedure,
-      level_1_file_path = level_1_path,
-      level_2_file_path = level_2_path
+      olink_assay_file = olink_assay
     ),
     quiet = FALSE
   )